New Job APIs (#2791)
This PR introduces new set of RESTful APIs to manage and query jobs, and
a new client to work with those APIs.

### Curl Usage Examples
#### Submit Job
```
curl -X PUT \
     -H "Content-Type: application/json" \
     -d '{
          "Job": {
            "Name": "test-job",
            "Namespace": "default",
            "Type": "batch",
            "Count": 1,
            "Labels": {
              "foo": "bar",
              "env": "dev"
            },
            "Tasks": [
              {
                "Name": "task1",
                "Engine": {
                  "Type": "docker",
                  "Params": {
                    "Image": "ubuntu:latest",
                    "Entrypoint": [
                      "echo",
                      "hello"
                    ]
                  }
                },
                "Publisher": {
                  "Type": "noop",
                  "Params": {}
                },
                "ResourcesConfig": {
                  "CPU": "0.1",
                  "Memory": "10mb"
                },
                "Network": {
                  "Type": "None"
                },
                "Timeouts": {
                  "ExecutionTimeout": 30
                }
              }
            ]
          }
        }' \
     http://0.0.0.0:20000/api/v1/orchestrator/jobs

{"JobID":"28c08f7f-6fb0-48ed-912d-a2cb6c3a4f3a","EvaluationID":"996b12e4-bcc5-4d74-ac21-0c421dafb7de"}
```

#### Query Job
```
# Get job
curl 0.0.0.0:20000/api/v1/orchestrator/jobs/28c08f7f-6fb0-48ed-912d-a2cb6c3a4f3a

# Get job (pretty output)
curl 0.0.0.0:20000/api/v1/orchestrator/jobs/28c08f7f-6fb0-48ed-912d-a2cb6c3a4f3a\?pretty

# List all jobs
curl 0.0.0.0:20000/api/v1/orchestrator/jobs

# List with options
curl "0.0.0.0:20000/api/v1/orchestrator/jobs?limit=2&reverse=true&order_by=CreateTime"

# List with label filtering
curl --get "0.0.0.0:20000/api/v1/orchestrator/jobs?pretty" --data-urlencode 'labels=env in (prod,dev)'

# List job history
curl 0.0.0.0:20000/api/v1/orchestrator/jobs/28c08f7f-6fb0-48ed-912d-a2cb6c3a4f3a/history

# List job executions
curl 0.0.0.0:20000/api/v1/orchestrator/jobs/28c08f7f-6fb0-48ed-912d-a2cb6c3a4f3a/executions

# List job results
curl 0.0.0.0:20000/api/v1/orchestrator/jobs/28c08f7f-6fb0-48ed-912d-a2cb6c3a4f3a/results

# Stop Job
curl -X DELETE 0.0.0.0:20000/api/v1/orchestrator/jobs/28c08f7f-6fb0-48ed-912d-a2cb6c3a4f3a
```

#### Query Nodes
```
# Get node by ID
curl 0.0.0.0:20000/api/v1/orchestrator/nodes/QmPc73cWi6d8trhEhastiVgypwDUvmK1UkzAX9pQgrGAg3

# List all nodes
curl "0.0.0.0:20000/api/v1/orchestrator/nodes?pretty"

# Filter nodes using label filtering
curl --get "0.0.0.0:20000/api/v1/orchestrator/nodes?pretty" --data-urlencode 'labels=name in (node-0,node-1)'
```

#### Query agent (current node)
```
# Is alive
curl 0.0.0.0:20000/api/v1/agent/alive

# Deployed version
curl 0.0.0.0:20000/api/v1/agent/version

# Node info
curl 0.0.0.0:20000/api/v1/agent/node
```

### Client Usage Examples
```
client := New(options{}, optFns)

# submit job
putJobResponse, err := client.Jobs().Put(&apimodels.PutJobRequest{
    Job: &models.Job{...},
})

# get job
getJobResponse, err := client.Jobs().Get(&apimodels.GetJobRequest{
    JobID: putJobResponse.JobID,
})

# agent version
versionResponse, err := client.Agent().Version()

# list all nodes
listNodesResponse, err := client.Nodes().List(&apimodels.ListNodesRequest{})
```

### Notes
1. BasicAuth has been added as an example on what to include in the
client options, and how they can be injected in each outgoing request,
but they are not checked in the server side yet
2. Request signing has been removed as we need to find a better
alternative to our previous approach to simplify client and sdk creation

ENG-55
ENG-56
ENG-58
ENG-59

+++ pkg/jobstore/boltdb/store.go
@@ -391,9 +391,9 @@ func (b *BoltJobStore) getJobsBuildList(tx *bolt.Tx, jobSet map[string]struct{},
 func (b *BoltJobStore) getJobsWithinLimit(jobs []models.Job, query jobstore.JobQuery) []models.Job {
 	limit := query.Limit
 	if limit == 0 {
-		limit = len(jobs)
+		limit = uint32(len(jobs))
 	} else {
-		limit = math.Min(len(jobs), limit+query.Offset)
+		limit = math.Min(uint32(len(jobs)), limit+query.Offset)
 	}
 
 	return jobs[query.Offset:limit]
